Add support for literal indexes beyond than 65 535
The SetLiteral instruction can only be used to access a maximum of 65
535 literals. That many literals in a single block is unlikely. However,
when we start storing literals per module (instead of per block),
reaching this limit is less likely.

In this commit we introduce the SetLiteralWide instruction. This
instruction can address up to 42 949 67 295 literals per module; which
should be more than enough. This instruction works by storing the higher
16 bits as the second argument, and the lower 16 bits as the third
argument.

SetLiteralWide is a separate instruction so the additional overhead of
bitwise operations doesn't affect code that doesn't need to address this
many literals.

+++ vm/src/bytecode_parser.rs
@@ -269,7 +269,7 @@ fn read_instruction<T: Read>(
 fn read_instructions<T: Read>(
     bytes: &mut Bytes<T>,
 ) -> ParserResult<Vec<Instruction>> {
-    let amount = read_u64_with_limit(bytes, u16::MAX as u64)? as usize;
+    let amount = read_u64_with_limit(bytes, u32::MAX as u64)? as usize;
     let mut buff: Vec<Instruction> = Vec::with_capacity(amount);
 
     for _ in 0..amount {
@@ -316,7 +316,7 @@ fn read_literals_vector<T: Read>(
     state: &RcState,
     bytes: &mut Bytes<T>,
 ) -> ParserResult<Vec<ObjectPointer>> {
-    let amount = read_u64_with_limit(bytes, u16::MAX as u64)?;
+    let amount = read_u64_with_limit(bytes, u32::MAX as u64)?;
     let mut buff = Vec::with_capacity(amount as usize);
 
     for _ in 0..amount {
